Extract ps4 addon
