+++ www/src/Locales/en/AddonsConfig.jsx
@@ -108,8 +108,10 @@ export default {
 	'socd-cleaning-mode-selection-slider-mode-two-label': 'SOCD Slider Mode Two',
 	'socd-cleaning-mode-selection-slider-pin-two-label': 'Pin Two',
 	'ps4-mode-header-text': 'PS4 Mode',
-	'ps4-mode-sub-header-text':
-		'<0>!!!! DISCLAIMER: GP2040-CE WILL NEVER SUPPLY THESE FILES !!!!</0> <1>Please upload the 3 required files and click the "Verify & Save" button to use PS4 Mode.</1>',
+	'ps4-mode-sub-header': 'DISCLAIMER!',
+	'ps4-mode-sub-header-text': 'GP2040-CE WILL NEVER SUPPLY THESE FILES!!!',
+	'ps4-mode-text':
+		'Please upload the 3 required files and click the "Verify & Save" button to use PS4 Mode.',
 	'ps4-mode-private-key-label': 'Private Key (PEM)',
 	'ps4-mode-serial-number-label': 'Serial Number (16 Bytes in Hex Ascii)',
 	'ps4-mode-signature-label': 'Signature (256 Bytes in Binary)',
